requirements.txt
requests
python-dotenv
openai
transformers
torch
